Add fsfs lite implementation
